requirements.txt
Django==5.0.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-tailwind==0.5.0
reportlab==4.0.9
xhtml2pdf==0.2.13